bug 75482 - work around incorrect options table layout in IE8
IE8 apparently gets the colspace=2 on the delegate options table
wrong. In order to work around this, we add a spacer cell so that the
table renders in to the second column. This doesn't look terribly
good, but it's much better than the previous buggy layout.

To compensate for the buggy layout, we include a slight style tweak
that makes the buttons align with the table above it.

http://bugzilla.zimbra.com/show_bug.cgi?id=75482