Bug 82732
save copy of send-as message to sent-as user's Sent folder

Fixed according to the description of this bug by Jeff.

First, I thought a simple fix in ZmComposeController.prototype._sendMsg  (after a lot of investigation) seems to work - set the accountName to the "send as" or "obo" which in both cases is set in msg.delegatedSenderAddr.

However, then the OBO would appear (i.e. the "s" type "from", meaning "sender" was included in the request), even when just doing "send as" and not OBO. I solved this in the weird long code of ZmMailMsg.prototype._addFrom, by fixing the isPrimary (though not directly affected by my fix in ZmComposeController.prototype._sendMsg)

Anyway as a result of those fixes, the sent messages appear in both user1's and user2's  "sent" folder for both user1 as user2 and user1 OBO user2 cases. This I believe is the requirement.

http://bugzilla.zimbra.com/show_bug.cgi?id=82732