bug:84178
Non-Contact items in Contacts folder cause Mac addressbook to report:
    CoreDAVHTTPStatusErrorDomain Code=404 The operation couldn?t be completed.
This fix makes the addressbook-multiget report succeed which means that the addressbook now loads, which is progress.
However, still getting Warning triangle with error "Some server contacts could not be copied to this address book", so, further work is needed.  Plan to address that as part of the fix for Bug 84331.

Enhanced logging in DavServlet for headers in request and tracking sendError on the response.
Also enhanced Log to support logging where the level is specified as an arg - punching through to the log4j equivalents.
https://reviewboard.eng.vmware.com/r/526355/
IM --> main

http://bugzilla.zimbra.com/show_bug.cgi?id=84178