Bug 84302
Saved search appears under both Searches and Folders

I saw bug 27483 for example as some of the origin of this code and logic, related to calendars perhaps. It seems to work now for search, but hopefully it's not fragile.

The fix is to only allow ("allow" here means render in the current tree view) folders if they are of the allowed type, or if they are of the allowed sub types (e.g. "search" that is allowed as sub type of a regular folder)... but ONLY if they have an ancestor that is of a "Folder" (allowedTypes) type that is NOT the root (descendent of the root).

e.g.
allowed:
Folders-->folder1--->searchFolder1
Folders--->folder1--->folder2--->folder3--->searchFolder1

not allowed:
Folders-->searchFolder1
Folders-->searchFolder1--->searchFolder2

I do that by recursion going up the ancestor tree.

I also removed the conditional code relying on this._allowedTypes, as this was never set (I can't find it in code anywhere at least now)... as opposed to this.allowedTypes.

http://bugzilla.zimbra.com/show_bug.cgi?id=84302