bug: 83070, 84346 - prevent IE from going into compatiblity mode
It appears that IE9 continues to render certain pages in quirks mode
in the QA environment. This may be because IE is configured to use
Compatibility Mode for Intranet sites, or some other similar cause.

The most certain way to prevent IE from going into quirks mode is
setting the X-UA-Compatible HTTP header. This only leaves two ways
which Zimbra might end up in compatibility mode and quirks mode:

* IFRAMEing Zimbra within a page rendered in quirks mode.
* A Group Policy forces the Zimbra domain into quirks mode.

Both of these should be considered advanced usecases, and will have to
be addressed by the customer.

While at it, I fixed a few more JSP files to use standards mode, and
moved the DOCTYPE up earlier in the element when relevant.

http://bugzilla.zimbra.com/show_bug.cgi?id=83070
http://bugzilla.zimbra.com/show_bug.cgi?id=84346