Bug 77882
contact group list is not populated with new groups

related to bug 81055. The root cause is that the canonical list is not updated in case of update from server notification (different client/tab/browser/etc)
This happens since the update of the canonical list is done in ZmContact.prototype._handleResponseCreate and that is called only as a callback from  ZmContact.prototype.create, and is not called in case of update from server. (notification). It's only called when user edits a new contact.

Solution (that might solve other issues other than this bug too) is to update the canonical list from ZmContact.createFromDom as well.

http://bugzilla.zimbra.com/show_bug.cgi?id=77882