Bug 84484
Button "Location" is disabled in shared address book

This is due to bug 40338. No idea why it was decided there that to fix it, not allowing changing the folder for shared folder.

So I allow changing the folder if the contact is not read only. So for a share with manager or admin right, it can change the folder. To make it work I also had to fix the part in ZmEditContactView.prototype.getModifiedAttrs  that was always adding the folderId if it's not shared (it was supposed to do it if it's NEW, as the comment was mentioning, but it wasn't, so I fixed that).

http://bugzilla.zimbra.com/show_bug.cgi?id=84484