bug 75492 - overhaul SMS Notifications preferences
The reported bug was that an alignment was off, and that this only
occurred in IE. However, I also saw this behaviour in Safari.

This change adjusts the layout of the SMS Notifications preferences to
be more symmetric as well, ensuring that buttons and labels align
vertically, and that the input fields are large enough that they reach
the buttons in all browsers.

http://bugzilla.zimbra.com/show_bug.cgi?id=75492