Bug 76734
Some issues with Distribution List context menu

This is a bit of a refactoring, to fix the case of editing a DL from the bubble. First, note that the bubble "knows" (on right click) that this is an editable DL only if the user went to the Distribution Lists folder under Contacts. I don't fix that for now.

The problem though was that the DL was not fully loaded with stuff we get from GetDistributionListRequest when coming from the bubble.

I wanted to refactor the DL Info gathering code, so it would not be called from ZmContactListController.prototype._editListener, but instead from ZmContactController.prototype.show (because show() is called from both cases). But for that, I realized that since we had all those methods in ZmContactListController, they would not be easily accessible from there when called from show() so I decided the easiest would be to move those methods to ZmContact.  I hope it's not too "controllery" stuff that's moved to model. On the other hand, all those methods access the contact methods now from "this" which is more natural. It all starts with ZmContactListController.prototype.gatherContactExtraDlStuff (now called gatherExtraDlStuff as it's a method ON contact) moved to ZmContact.prototype.gatherExtraDlStuff. And a bunch of methods are are called by it etc.

http://bugzilla.zimbra.com/show_bug.cgi?id=76734