bug: 83261
Was getting error that can't change ORGANIZER from "a" to "a".  i.e. the from and to values were the same.
Don't have a full reproduction scenario but as Invite's "mIsOrganizer" is explicitly set via many different code paths and is not re-computed it is perfectly possible that there are some code paths where it isn't updated correctly.
The fix here is to assume that if an address was OK for the original invite, it must be OK for the new one too.
IM ---> main
https://reviewboard.eng.vmware.com/r/535040/

http://bugzilla.zimbra.com/show_bug.cgi?id=83261