bug 78163 - add contact images to message views
The image is rendered to the left of the subject and to/from
information. We use the standard contact image functions, and offer no
special mouse-over effects at this time.

We use different image sizes in different locations, depending on what
fits: The normal message view have 48x96; expanded conversation
messages have 48x48 and collapsed conversation messages have 32x32.

One possible issue is that the images render with a visible delay on
every load. This is caused by the server forcibly disabling cache on
contact images.

http://bugzilla.zimbra.com/show_bug.cgi?id=78163