bug: 79576
Clicking on a tag which isn't defined for the mailbox of a shared folder results in shared items with OTHER tags being returned in the results.

Fix is to do an optimization step which affects the query against the mailbox for the shared folder to prune out Intersections where one of the terms is a check for a tag which isn't defined for that mailbox.
https://reviewboard.eng.vmware.com/r/536119/
IM to main + added test

http://bugzilla.zimbra.com/show_bug.cgi?id=79576