bug: none - remove unused and broken Dwt.toggle()
Judging from the arguments, Dwt.toggle(elem, false) would hide an
element. That's not the case, however, as false compares equal to
null. To make matters worse, the function neglects to declare its
local variable (appropriately named “it”) which it thus leaks to the
global context.

The function has no users -- I found no references in ZimbraWebClient,
in Zimlet nor in ZimbraNetwork -- leaving little value in actually
fixing or improving it.