bug 81717 - use TinyMCE 4.0 for the rich text editor
This deactivates all current Zimbra-provided plugins and
customizations, as TinyMCE's internal API has changed
substantially. For the one important feature of allowing direct image
uploads, we roll a new, simpler plugin which is contained within
ZmAdvancedHtmlEditor. For breaking BLOCKQUOTEs on enter, we extend the
keydown handler.

Two issues remain: Emoticons don't work, and the font style and size
aren't updated in the toolbar.

Significant changes to TinyMCE include a completely new skin, which
incidentally integrates nicely with the default 'serenity' skin. I've
applied a few CSS tweaks to make TinyMCE use the same fonts in buttons
and menus as the rest of the AJAX UI.

Toolbars reflow nicely in TinyMCE 4 as the window gets progressively
smaller; as such, there's no need for a second, hidable
toolbar. Instead, we include the elements of the second toolbar in the
first. I'll fix up the order to be more logical in a follow-up change.

The current tweaks to shortcuts no longer work, so we remove them. If
necessary, I will address this in a follow up change as well.

TinyMCE 4 fails to load if you specify a non-existant locale, so we
generate a list of supported locales at build time, and use it for
checking the Zimbra locale.

All previous styling tweaks no longer apply to the new skin, so
editor_ui.css is gone. TinyMCE 4 doesn't support specifying an editor
CSS script anyway.

We also apply a slight refactoring to the file names in the TinyMCE 4
distributable, and replace the dot as an internal separator with an
underscore. Otherwise, AjxPackage would fail to load the files.

http://bugzilla.zimbra.com/show_bug.cgi?id=81717