Bug 80785
"undo" delete/move should not place user at top of the list


A bunch of things collaborated to form this bug. I fixed the following things to make it work:

1. the list selection was lost due to "hijack" of the savedState before sorting (we need to sort since we don't know where the item has to be in the list - but see a TODO at the end of this) since restoreState was called (by something related to keepReading in resetOperations somewhere) before the search completed and the callback (restoreState) had a chance to be called. Thus the savedState was emptied.  To solve it I now allow a param in the restoreState which is the state you want to restore to, and I pass that param in to the restoreState in the callback closure for what used to be sortColum (more on that in #2)

2. unfortunately, the reading pane view is not restored on restoreState - that had a whole different issue. ZmDoublePaneController.prototype.show was calling reset on the itemView, because that's what it does since bug 76315 (actually before too) if the search does not have a "isRefresh" attribute, or the view does not include the search selectedItem attribute.  To fix that I eventually did two thing:

2a. I noticed that if sort is needed (if an item was moved to current list) in ZmListView.prototype._changeListener, it was calling sortColumn - this was more complicated to set up I believe than simply doing a redoSearch which is explicitly what we need to do. So I changed it to do that.
2b. in my new _redoSearch in ZmListView I set a new param isRedo that is so far used exactly as isRefresh in ZmDoublePaneController.prototype.show to determine we might not need to reset the item view. I also added the selectedItem as a param (change) to the getSearchController().redoSearch call. And the callback passed include the param state (passed as this._state) as I mentioned in item (1).

3. Another bug I fixed which is also happening in the case of refresh block (can be emulated with $set:refresh in the search bar) is that the toolbar was all disabled after the refresh. I fixed that by calling resetOperations correctly in case the selected item was available. (it was lost in the call to this._setup in ZmDoublePaneController.prototype.show)

This now works and the focus from the next item chosen is not lost (or rather, regained without the user noticing) when undoing a previous item move or delete.

A very nice TODO would be to get the sort index like we do in notifyCreate, in case the sort is ascending or descending by DATE. We should be able to re-use the ZmMailList.prototype._getSortIndex to get the index in that case and insert it there, instead of calling the server redoSearch.

P.S. - fix 1 might not be needed, I came to it before fix 2... but I think the part in 2 fixes 1 too. However, I don't think it hurts to keep fix 1 as those saved states are very fragile. Might be a good idea to always include that state param in the callbacks to restoreState, in other cases too. Oh well...

http://bugzilla.zimbra.com/show_bug.cgi?id=80785