bug 65747 - do last-ditch save even for 'invalid' tasks and appointments
We bypass most of the validity checking logic, and simply save the
task or appointment as-is rather than inserting dummy values. As a
consequence, the user must enter these values next time they open the
calendar item; otherwise they won't be able to save their changes.

One might argue that having an incomplete calendar item is a bug, or
that the behaviour described above is the closest possible match to
the intent of the user, and thus a feature. That is not the primary
motivation here, however. Our main concern is saving the data NOW, no
matter what.

Judging from the code, we have many more ways a calendar item might be
'invalid' other than just missing a summary. As this code path is
unlikely to receive much testing, treating all cases the same
increases the likelyhood that the code actually works, and continues
to do so.

http://bugzilla.zimbra.com/show_bug.cgi?id=65747