Bug 78238
Time/Date/Weekday highlight in Webmail needs option to disable;  "So" (Sonntag) too common

Since the code relies on everything being case Insensitive, the only feasible solution to allow the "so" to now be insensitive to case was to allow on a per-language to set the format to be case Sensitive.

For that I added a format.caseSensitive key in the zimlet's properties file. by default if it's not there, the case is INsensitive. If it's there and equal to "true" (or rather boundaryTrue, which is for some reason externalized for the boundary boolean, so I re-use it to be consistent), then the regexp is not case sensitive.

One thing though that happens when making it case sensitive is that all the formats are case sensitive. I thought that might be a problem in the "ordinal" case (e.g. Third Thursday in April), so for German I changed the ordinal properties to include both lowercase and uppercase first letter. e.g.:
calc.ordinal.first = 1.|erster|Erster

It's not perfect but that's the solution I have.

http://bugzilla.zimbra.com/show_bug.cgi?id=78238