Bug 84924
GC: Can't put focus in Cc/Bcc textarea and text cursor is invisible

regression from bug 84328.

The first time ZmAddressInputField.prototype._resizeInput is called for the CC and BCC elements, the _holder is not yet sized or something (Dwt.getSize(this._holder).x is 0) and maxWidth becomes -1 or so.

My solution is maybe a little hacky (maybe we should not call this when the _holder does not yet have a size), but it works, make sure we don't go below 3 (a little arbitrary, maybe 1 would have worked too) in the input size, so the cursor WILL appear. Once the user starts typing, the resizes work better anyway.

http://bugzilla.zimbra.com/show_bug.cgi?id=84924