bug 85043 - unbreak production builds on Safari
From some reason or other, Safari fails to resolve the variable 'i' on
line 112 of ‘ZmMailApp.js’. I don't see any particular reason why the
code in question should fail, but adding ‘var’ to the for loop
declaration fixes it.

My best guess is that the root cause either lies in Safari or our
minifier.

http://bugzilla.zimbra.com/show_bug.cgi?id=85043