Bug 78388
The contact without any value is saved by keyboard shortcut "Ctrl+S"

I think the straight forward fix is to check whether the "Save" button is enabled from the keyboard handler. That way any logic that changes would be reflected here automatically. If the Save button is not enabled, do nothing with CTRL-S.

Took this opportunity to use the ZmBaseController getCurrentToolbar method instead of repeating the same code, in multiple places in this file.

http://bugzilla.zimbra.com/show_bug.cgi?id=78388