bug 81717, 84907 - update standard html client editor to TinyMCE 4.0
The compose view and signature editor in the standard client use
TinyMCE in HTML mode, and were broken by the change to the AJAX UI
that upgraded TinyMCE to 4.0.

This change changes the two tag files to use TinyMCE 4 event handlers,
configuration and method calls. In addition, it consolidates
conversion of Zimbra locale to TinyMCE language into a separate file
shared by the AJAX and Standard clients. Finally, we optimize TinyMCE
loading using a new tag file that provides the minimized source code
inline in the main document.

In addition, we alter our emoticon plugin to no longer use Ajx*
classes.

http://bugzilla.zimbra.com/show_bug.cgi?id=81717
http://bugzilla.zimbra.com/show_bug.cgi?id=84907