Bug 77930
Need unique/predictable ID for "File As" menu

Solution is not ideal. At first I changed the constants ZmContact.FA_LAST_C_FIRST etc from 1,2,... to
meaningful values, but then this was what was passed to the server and the server accepts the constants as 1,2,3. (you would think a string, even such as "LF" (for last, first) would be better but oh well).
So I decided to just pass explicit IDs that are meaningful to the options.

That's where things got weirder, since turns out there was not yet support of passed in ID in DwtSelectOption (only DwtSelectOptionData, recently, bug 79554), and the createMenu was just using the value as ID part... etc. Hard to explain.  Maybe it's not really needed as this case that the value is not meaningful (1, 2) is due to the server expecting this. But I added the support for the explicit ID anyway.

Also, the ID WAS set to the pseudoitems (which are created to get the correct size of the button in case it needs to match the largest item width). This is since recently - Bug 79554. But I don't think the ID should really be there, but in the real visible (not the pseudo) menu elements. I remove this ID, I hope it won't break tests but probably not many if at all, since it's new, and I think my solution is a little better.

Bottom line is the IDs are now as FA_COMPANY_LAST_C_FIRST_1, etc.

http://bugzilla.zimbra.com/show_bug.cgi?id=77930