Bug 82491
Need unique/predictable ID's on Right Click -> Group By menu

There was already attempt to have ID for the "Group By" element, but ZmId.GROUPBY was not defined.

I now use an inline constant instead, so now the ids are such as "GROUP_BY_1" (allowing for multiple mail views, such as search views with no ID clashes).

The ID of the entire menu is now "SORT_MENU_1" (etc).

The other sort options already had the ID from the short value... but that's probably why this bug was not complaining about this. :)

http://bugzilla.zimbra.com/show_bug.cgi?id=82491