Bug 58505
Meeting with only location shows "Attendees" section in hover

Problem is otherAttendees in the ZmAppt is set to true in case other attendees is just location. Set the already existent (for other reason, not sure what use-case) hideAttendees template param, to true/false based on whether there are "human" attendees.

http://bugzilla.zimbra.com/show_bug.cgi?id=58505