bug 84887 - add utility methods for calculating insets & bounds
Dwt.js has getInsets() and getBounds() as well as a method for
combining them; insetBounds().

DwtControl, on the other hand, has no support for calculating insets
directly, requiring using the HTML element directly instead.

All of this seems odd to me, and I'd like to use them in the fix for
bug 84887. As a result, this change adds a getInsetBounds() method to
Dwt and DwtControl, and a getInsets() to DwtControl.

http://bugzilla.zimbra.com/show_bug.cgi?id=84887