bug:76968
Support "fullConversation" attribute for SearchRequest.
If set to true, then for conversation hits, ALL messages in the conversation are included in the results.  The default is that only messages which match the search criteria are included.
Note that even junk and trash messages are included.  If a client switches to using "fullConversation" it may need to change handling of such messages - in my experiments with ZWC, when expanding conversations, deleted messages no longer have strike-through.

Also, use JAXB for processing the request elements of Search/SearchConv and SearchMultipleMailboxes
https://reviewboard.eng.vmware.com/r/548804/

http://bugzilla.zimbra.com/show_bug.cgi?id=76968