Bug 77796
Name of a new contact shown in auto complete field does not consider the value of FileAs

I find it weird that we don't set the fullname on create, and only on modify, no idea if we ever did.. but oh well, this is a fix.

btw, contrary to comment #5, it DOES include the fileAs... (i.e. the code for how we file, not the actual string that we files as, as that is the fullname attribute. Terminology is a bit confusing).

http://bugzilla.zimbra.com/show_bug.cgi?id=77796