Bug 58292
Enable/disable menus while multiple appointments are selected in list view

Implemented the request here exactly as requested. After investigating further, I now understand we can try to allow "delete", "move" and "tag" but disable everything else for multi selections.  The "delete", "move" and "tag" are also the actions in the toolbar, and allow for multi selected actions, so it makes sense.

I think this is mostly a visual change, and also avoids confusion when users previously thought actions such as "accept" applied to all items selected. The reason we can't do multi actions on other actions, is that there are just too many permutations on the following attributes at least:
1. Is it recurring?
2. Is the current user the organizer or just an invitee?
3. Is it accepted or declined or tentative?

I think to keep things simple this solution is fine.  It allows multi action on most cases for delete, move and tag.

http://bugzilla.zimbra.com/show_bug.cgi?id=58292