Bug 74373 : Color: dropdown control width is un-necessarily broad Bug 83521 : "Include:" dropdown row along over clickable in "Add Shared Folder" dialog Bug 83620 : Clicking on empty area near 'Task list' or 'Status' field ,respective dropdown list gets open
The Drop down widget is placed in div which has a width auto which means that it occupies the entire space in which it is placed. The table elements within that div are correctly sized. The fix is to add display:inline-block to the wrapping div so that wdith of div is same as the width of table within it. I am afraid to do this style change at the widget level as this widget is used all over the place and in Admin console too.

http://bugzilla.zimbra.com/show_bug.cgi?id=74373