Bug 77208
Child folder shown as parent folder for shared folders in "Find Shares" dialog

Went for the simple solution of showing the slashes so the user sees it's a path. And the actual name for the mountpoint would be the same as the default one for the "Accept" from the email message name, which is the owner name and the name of the folder (without the path leading to it). So there will be no slashes.

bug 64189 removed the slashes, since the name displayed was used as the name of the mount point, yet slashes are illegal in it.

So I un-removed the slashes (except for the first one that is always the first character), and removed them only in the name the mountpoint would be called. Made some other cleanups.

http://bugzilla.zimbra.com/show_bug.cgi?id=77208