bug 85132 - always display current font name & size in the editor
The main sticking point compared to TinyMCE 3 is how to get the
objects corresponding to each toolbar button control. Previously,
you'd simply query the controlManager for a control with a given
name. However, TinyMCE 4 removed the control manager with the
following justification:

"Each control in TinyMCE 4 is supposed to be an self contained
unit. This enables controls to be created at multiple locations in the
UI. So the old controlManager with fixed names has been removed."

http://www.tinymce.com/wiki.php/Tutorial:Migration_guide_from_3.x

Instead, we rely on a heuristic where we assume that controls are
unlikely to share both type and text.

In order to have the font size and text comparisons work across all
browsers, we normalize values to integer pixels or remove extraneous
quotes and spaces, respectively. This works in modern browsers as well
as IE8.

http://bugzilla.zimbra.com/show_bug.cgi?id=85132