bug 85058 - assign a width of 90% percent to newly created tables
In order to do this, we bind a listener to TinyMCE's BeforeSetContent
event, and match the output to what we expect from the table
plugin. The downside to this is that if someone were to e.g. paste a
table that happens to match what TinyMCE generates by default, we'd
resize it to 90%.

(I managed to work around this, but only by adding bindings to
detect whether the table button menu was visible. Bindings which I had
to insert from another listener, as the menu is created lazily. It
wasn't pretty.)

As the table button doesn't have a title, we switch to using
tooltip+type to 'identify' TinyMCE controls. Eventually, we'll
probably want to move to something more flexible.

http://bugzilla.zimbra.com/show_bug.cgi?id=85058