Bug: 81405 - iPad: Folder and tag management - create, move, rename, delete
Refactor and clean up folder management. Get it working
for contacts folders as well. During initialization, the
folder selector has a list of all folders, which is then
filtered before being displayed within an app’s overview
as a list of move targets.

- move controller/mail/ZtFolderListController to controller/ZtOverviewController
since it manages tags as well, and for multiple apps

- use generic selectors in config

http://bugzilla.zimbra.com/show_bug.cgi?id=81405