Bug: 82023 - On reply, conv moves to top
Fix and clean up handling of create notifications so that they
make proper use of the ability to match the current search
against a newly arrived msg or conv. A newly arrived conv will
be added to the top of the list if it matches. If a newly
arrived msg matches, its conv will be updated and bumped to the
top of the list.

- add “has:attachment” to the list of supported query matching
terms; if that’s the current search and a new conv with an
attachment arrives, it will show up (add to web client as well)

- fix double HTML-encoding of fragment from notification

http://bugzilla.zimbra.com/show_bug.cgi?id=82023