Bug 85316
User supplied search query should be parenthesized before augmenting with preference  terms

added parentheses around the query to make sure the entire query is "and"ed with the query hint (not sure why that part is called a query hint).

Took this opportunity to extract small piece of code that was duplicated between ZmSearch.prototype._getStandardMethodJson and ZmSearch.prototype._getStandardMethod, to get the query.

http://bugzilla.zimbra.com/show_bug.cgi?id=85316