Bug 85558
After user unchecking all Zimlets in preferences, and reloading - the zimlets disappear from prefs

Regression from Bug 60559 + another issue maybe not related.

It was calling getZimlets() of the ZmZimletMgr, which returns only the active zimlets. And used that to determine if there are zimlets. That's a bug, it should use the same zimlets list it later displays, that includes all the zimlets, which is obtained by ZmZimletsPage.getZimlets (it's a bit confusing since I renamed some stuff and refactored some repeated code).

The other issue after fixing that, was that it was not loading the Zimlet codes (which is essential to getting the i18n stuff working displaying the label/description/etc in the prefs... (without this fix it was showing stuff like "${....label}".

I don't know why that code was stopping in case of 0 zimlets that are active but I fixed that.

http://bugzilla.zimbra.com/show_bug.cgi?id=85558