Bug 81939
Need unique/predictable IDs on Contact Group headers

The original bug is for the predictable IDs but I noticed a regression from my fix (mainly from my refactoring for code cleanup).

Before the fix it was not showing the "location" on the contact split view for DLs.

After the fix it was showing "Location:" label but without the location.

What I decided to do is actually show the Location: Global Address List
That's consistent with the other GAL items. At first I did Location: Distribution Lists, but that is not always true, if this DL is a result of a GAL search but the user is not the owner or a member of.

Made some code cleanup too, we got the folder/address book twice within a few lines of code in 2 different ways, so I removed that.

http://bugzilla.zimbra.com/show_bug.cgi?id=81939