Bug 76771 - Script error XMLHttpRequest Exception 101 after waking from standby
The steps to reproduce this are still unknown but looking at the code and stack trace it seems that the reminder code was getting invoked when the system came online and the network connection was still not ready. All the reported instances of this exception point to the same stack trace so as a workaround added a try catch block surrounding the getRefreshParams method. If there is a exception then try to run that code again after 1 sec for a maximum of 3 attempts.

http://bugzilla.zimbra.com/show_bug.cgi?id=76771