bug 85384 - add methods for calculating insets for a child element
Zimbra tends to have margins and paddings all over the place. The
previous approach has generally been to insert magic constants -- say
add a 5 to a size here, subtract 10 there, and so on. Such constants
are notoriously fragile, and may even vary depending on the browser or
even skin. This can result in sizes eventually becoming subtly wrong,
such as an element having a padding to the left, but not the right.

Instead, we add a method to DwtControl for calculating the insets for
a child, along with a couple of helper methods in Dwt and AjxUtil. The
first user is ZmHtmlEditor.

(AjxUtil.reduce() isn't required, strictly speaking, but this an
excellent for introducing it, as I've missed it more than once, and it
makes this particular code much simpler.)

http://bugzilla.zimbra.com/show_bug.cgi?id=85384