bug 85384 - fix appointment editor resizing
This should ensure that the editor resizes properly, regardless of
whether it's caused by a window resize or toggling one of the other
controls.

The only remaining aesthetic issue I can see is that the appointment
subject is slightly too large, caused by the combination of a 100%
width and a padding.

http://bugzilla.zimbra.com/show_bug.cgi?id=85384