Bug 85455
Create
Appointment/Task/New Filter action doesn't switch focus to model dialog or zimbra main window

Simply focusing on the opener window is either hard/hacky or impossible. See here:  http://stackoverflow.com/questions/18501095/window-opener-focus-doesnt-work

I tried to add a call to opener.window.focus() from ZmMailListController.prototype._handleResponseFilterListener in case of new window, but that didn't work. (The focus on the opener window is correctly in the name field of the new filter, but only after I close the child window).

My solution is simple - close the child window when the user does a "Create filter/task/appt". Since the actions are from a "view only" child window (not compose for example), nothing can be lost. It may not be ideal but even if I was able to focus back to the parent window, there's a problem with that - the child window would be hidden and the user might not even know it's open behind the parent window, which is also a bad practice.

http://bugzilla.zimbra.com/show_bug.cgi?id=85455