Bug 85773
Script Error: Cannot call method 'setHasMore' of null

The exception is in ZmListController.prototype.setHasMore, because this._list is undefined or null.

Added a null check for this._list although the calling code is weird, I don't know why it gets the controller on which it calls seatHasMore in a different way, even though ZmMailApp.prototype._checkList is passed a controller (in a loop of controllers from ZmMailApp.prototype.createNotify).

http://bugzilla.zimbra.com/show_bug.cgi?id=85773