bug 85470 - fix signature editor resizing
This fixes horizontal resizing of the signature editor, which
otherwise had trouble acquiring the sizes of the containing elements.

Two resize handlers is one too many for ZmSignaturesPage, so fold them
into one. To be on the safe side, we apply a slight tweak to
ZmHtmlEditor, which makes us always apply sizes to text field &
editor, when possible.

http://bugzilla.zimbra.com/show_bug.cgi?id=85470