Bug: 76597   "Invitees:" header shows blank in auto-accecpted respond email
The invite auto-reply from the location only contains itself as an invitee.  Talking with Josh, we agreed that we will suppress the 'Invitees' label if its an invite, but there are no visible invitees.  The user can still easily get the invitees from the first message in the conversation (the original invite) or by going to the appt.

http://bugzilla.zimbra.com/show_bug.cgi?id=76597