Bug 85657
[ZCS7.1.3.j.c.1] Not load the save search list for appointment

We had a mess with the searchFor variable. Still a little messy. The problem was in the case of saved search, the "types" were not set and so we used the searchFor - but that in the case we click on the Calendar saved search from Mail (or Contacts) was probably wrong (as it was probably taken from the search bar, which is not related to the search we choose).

To solve it what I do is always set the searchFor to the matching value based on the types (if we have types... if we don't get types, we set them based on searchFor first... confusing, I know).

I simplified the getTypes method to return just one type (as it did, but it wrapped it in a vector), and renamed it to getTypeFromSearchFor, and added a matching getSearhForFromType.

Note the relationship between searchFor and type is pretty simple:
searchFor === type for everything except for mail.
searchFor = ZmId.SEARCH_MAIL for type of ZmItem.MSG || ZmItem.CONV

I also did some code cleanups to take advantage of only having a single type for a search.

http://bugzilla.zimbra.com/show_bug.cgi?id=85657