Bug 85399
WebEx zimlet fails for account

Regression from bug 70864, probably. There it set the Telephony/TSP, and if tspEnabled is false,it set telephonySupport in the request (via the callInType variable in the code) to "OTHER" (if TSP is enabled, which is more common probably, it set it to "CALLIN").  I found here: https://developer.cisco.com/media/webex-xml-api-11-05-2013/814SetSalesSession.html
That "NONE" is an option for telephonySupport. Not sure if OTHER means anything but NONE does seem to work

http://bugzilla.zimbra.com/show_bug.cgi?id=85399