Bug 85374
Sending mail with attachment as delegated user, attach a wrong file (usually an unrelated mail)

Qualify the "did" (draft ID, as this was also a problem, it would not delete the draft, or would delete a wrong unrelated message or draft if happens to have same ID), and also attachment mid's.

Part of the fix (in addition adding the qualifying code) was to do the qualification (for other parts too) not just in case of multiple accounts (not sure what this case covers exactly), but in any case the accountName is passed to _createMessageNode. (which happens in case of send as, and send on behalf of, which I also tested and it works now).

http://bugzilla.zimbra.com/show_bug.cgi?id=85374