Bug 85958
Not able to drag and drop contact to Trash

Regression from bug 4098, not from the core functionality of bug 4098 but from a rare case where using !== instead of != breaks stuff. Unfortunately nId is a String "3" while ZmFolder.ID_TRASH is a number 3, and they are not ===...

We thought about different solutions, changing the constants to be Strings is one of them. I thought of converting the localId in ZmOrganizer.normalizeId to Number. But anyway that would be left as a later issue.

For now it seems the whole treatment of contacts in ZmFolder where ZmFolder.prototype.mayContain checks that the target folder is Trash only, is not needed as there's no way for the user to try to move a contact to a non-contacts folder anymore. (used to be able when we had the mixed search view). So I removed that.

In the future, unless we have a better solution, try to remember not to change == to === and != to !== where folder IDs are involved, as it might break stuff.

http://bugzilla.zimbra.com/show_bug.cgi?id=85958