bug 84676 - suppress unnecessary warning about data loss for empty messages
This appears to be a regression caused by change 411689 which fixed
bug 78085 and bug 77707. As part of the general overhaul in that
change, the logic to check if the message was accidentally dropped. In
particular, the switchToText argument to _warnUserAboutChanges() in
ZmComposeController wasn't actually used.

Instead, we simply drop the argument, and inline the check for
converting a non-empty HTML message to text into the willLoseChanges
variable.

http://bugzilla.zimbra.com/show_bug.cgi?id=84676