bug:83753
Allow Delegated Administrators the right to set/modify owner of distribution lists.
There is a bootstrapping problem for assigning owners to DLs created by a domain admin.
Before this change, only a full admin or someone who already has ownDistList rights could assign owners to a DL.
This change allows domain admins who have createDistributionList (or createGroup) capability to assign owners to a DL or group.
Note that the default combo right domainAdminConsoleRights includes the createDistributionList capability but NOT the createGroup capability, so by default this change will only affect non-dynamic distribution lists.

Also corrected the spelling of "insufficient" in a couple of error messages.
Introduced an "=" between the target type and name in a couple of debug messages + used formatting instead of string concats.
https://reviewboard.eng.zimbra.com/r/116/
IM ---> main

http://bugzilla.zimbra.com/show_bug.cgi?id=83753