Bug: 85226  Icon not updating state after forward/reply
If a mail message had already been marked as read (viewed earlier and now revisited), when the user revisited the message, its notification info (ZmItem._evt.item and its list.item) were not properly set up.  Then, if the user did a 'Forward' or 'Reply' command for the mail message, ZmMailMgListView received the mailList as the item being notified (instead of the mail message), and so did not update the display of the affected mail message properly.

Also, the composeView was setting the isForwarded flag locally, so that the incoming notification was not believed to be a change.  So the normal notification system did not cause an update either.

http://bugzilla.zimbra.com/show_bug.cgi?id=85226