Bug: 85193 - Making it so stock browsers on Android will hide the address bar.  Chrome on Android will not hide the address bar no matter what, see:
https://developers.google.com/chrome/mobile/docs/faq?hl=ja

"Can I hide the URL bar?
The URL bar in Chrome for Android is Chrome’s Omnibox, with all of the built-in search and navigation features. For this initial release, we decided to show the Omnibox at all times. We are monitoring usage patterns and feedback to decide how to evolve this feature in future iterations of the application."

http://bugzilla.zimbra.com/show_bug.cgi?id=85193