Bug 85137
Clicking to search folder switches to different view rather than saving previous state

This was around the same area in code as bug 85657, and continues to be a bit tricky. I hope I simplified it a bit more.

I removed the part about _prevType that I do not believe was needed anymore.

I had to also make it update the user setting for the view, even if we change it in search results or in a view that is default to message view. This is good for consistency.

http://bugzilla.zimbra.com/show_bug.cgi?id=85137